Finished LIDAR feature
The LIDAR feature should be fully functionable now
